requirements.txt
flask
gunicorn
smartapi-python